refac(strategy-details): split the strategy details page into smaller independent components
